requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pandas-stubs>=2.0.0  # Type stubs for pandas

# Development tools
ipykernel>=6.25.0
jupyterlab>=4.0.0
notebook>=7.0.0

